Extracting some of the journal related code from calendar
